Minor code reorganization
